+++ changelog.md
@@ -4,6 +4,10 @@
 
 - Shaark is now a Web Share Target (only on Android [#65](https://github.com/MarceauKa/shaark/issues/65))
 
+## Fixed
+
+- WebParser can now return null values ([#66](https://github.com/MarceauKa/shaark/issues/66))
+
 ## Removed
 
 - Github actions